+++ clarify_python/constants.py
@@ -3,8 +3,7 @@
 __author__ = 'Paul Murphy'
 __email__ = 'paul@clarify.io'
 
-__version__ = '1.0.4'
+__version__ = '1.1.0'
 __api_version__ = 'v1'
 __api_lib_name__ = 'clarify_python'
 __host__ = 'api.clarify.io'
-__debug_level__ = 0  # Set to 1 if you want to see debug output from HTTP ops.